fix: interchange proxy names to match the direction of traffic
